refactor: migrate the `Guides` story to TypeScript
Ease maintenance, detect errors earlier.
The KeyHandler feature is not working with storybook, and some prototype elements are overridden. This will be handled later.
